// packages/platform/src/types/mt5/symbol/symbol.type.ts
export interface Mt5Session {
  sunday: Mt5SessionTime[];
  monday: Mt5SessionTime[];
  tuesday: Mt5SessionTime[];
  wednesday: Mt5SessionTime[];
  thursday: Mt5SessionTime[];
  friday: Mt5SessionTime[];
  saturday: Mt5SessionTime[];
}

export interface Mt5SessionTime {
  open: number;
  openHours: number;
  openMinutes: number;
  close: number;
  closeHours: number;
  closeMinutes: number;
}

export interface Mt5Symbol {
  sessionsQuote: Mt5Session;
  sessionsTrade: Mt5Session;
  symbol: string;
  path: string;
  isin: string;
  description: string;
  international: string;
  basis: string;
  source: string;
  page: string;
  currencyBase: string;
  currencyBaseDigits: number;
  currencyProfit: string;
  currencyProfitDigits: number;
  currencyMargin: string;
  currencyMarginDigits: number;
  color: number;
  colorBackground: number;
  digits: number;
  point: number;
  multiply: number;
  tickFlags: 'TICK_REALTIME';
  tickBookDepth: number;
  filterSoft: number;
  filterSoftTicks: number;
  filterHard: number;
  filterHardTicks: number;
  filterDiscard: number;
  filterSpreadMax: number;
  filterSpreadMin: number;
  tradeMode: 'TRADE_DISABLED';
  calcMode: 'TRADE_MODE_FOREX';
  execMode: 'EXECUTION_REQUEST';
  gtcMode: 'ORDERS_GTC';
  fillFlags: 'FILL_FLAGS_NONE';
  expirFlags: 'TIME_FLAGS_NONE';
  spread: number;
  spreadBalance: number;
  spreadDiff: number;
  spreadDiffBalance: number;
  tickValue: number;
  tickSize: number;
  contractSize: number;
  stopsLevel: number;
  freezeLevel: number;
  quotesTimeout: number;
  volumeMin: number;
  volumeMax: number;
  volumeStep: number;
  volumeLimit: number;
  marginFlags: 'MARGIN_FLAGS_NONE';
  marginInitial: number;
  marginMaintenance: number;
  marginLong: number;
  marginShort: number;
  marginLimit: number;
  marginStop: number;
  marginStopLimit: number;
  swapMode: number;
  swapLong: number;
  swapShort: number;
  swap3Day: number;
  timeStart: number;
  timeExpiration: number;
  reFlags: number;
  reTimeout: number;
  ieCheckMode: number;
  ieTimeout: number;
  ieSlipProfit: number;
  ieSlipLosing: number;
  ieVolumeMax: number;
  priceSettle: number;
  priceLimitMax: number;
  priceLimitMin: number;
  tradeFlags: 'TRADE_FLAGS_NONE';
  orderFlags: 'ORDER_FLAGS_NONE';
  optionsMode: 'OPTION_MODE_EUROPEAN_CALL';
  priceStrike: number;
  marginRateLiqudity: number;
  faceValue: number;
  accruedInterest: number;
  spliceType: 'SPLICE_NONE';
  spliceTimeType: 'SPLICE_TIME_EXPIRATION';
  spliceTimeDays: number;
  marginHedged: number;
  marginRateCurrency: number;
  filterGap: number;
  filterGapTicks: number;
  chartMode: 'CHART_MODE_BID_PRICE';
  ieFlags: number;
  volumeMinExt: number;
  volumeMaxExt: number;
  volumeStepExt: number;
  volumeLimitExt: number;
  ieVolumeMaxExt: number;
  category: string;
  exchange: string;
  cfi: string;
  sector: 'SECTOR_UNDEFINED';
  industry: 'INDUSTRY_UNDEFINED';
  country: string;
  subscriptionsDelay: number;
  marginRateInitialBuy: number;
  marginRateInitialSell: number;
  marginRateInitialBuyLimit: number;
  marginRateInitialSellLimit: number;
  marginRateInitialBuyStop: number;
  marginRateInitialSellStop: number;
  marginRateInitialBuyStopLimit: number;
  marginRateInitialSellStopLimit: number;
  marginRateMaintenanceBuy: number;
  marginRateMaintenanceSell: number;
  marginRateMaintenanceBuyLimit: number;
  marginRateMaintenanceSellLimit: number;
  marginRateMaintenanceBuyStop: number;
  marginRateMaintenanceSellStop: number;
  marginRateMaintenanceBuyStopLimit: number;
  marginRateMaintenanceSellStopLimit: number;
  volumeMinLots: number;
  volumeMaxLots: number;
  volumeStepLots: number;
  volumeLimitLots: number;
  ieVolumeMaxLots: number;
}